requirements.txt
Django==3.0.3
djangorestframework_jwt==1.11.0
djangorestframework==3.11.0
